compliance_checker/requirements.txt
# OpenAI Responses API Compliance Checker Dependencies

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# SSE (Server-Sent Events) parsing
sseclient-py>=1.8.0

# Configuration
pyyaml>=6.0.1

# JSON validation
jsonschema>=4.19.0

# CLI
click>=8.1.7
rich>=13.7.0  # For beautiful terminal output
tabulate>=0.9.0

# Reporting
jinja2>=3.1.2  # For HTML reports
markdown>=3.5.0

# Utilities
python-dateutil>=2.8.2
